Fix: 포스트 작성 버튼 로그인 예외처리 적용
